Refactored Profile and FeedItem components
ref https://linear.app/tryghost/issue/AP-286/outgoing-like-activities
